requirements.txt
numpy>=1.24.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pytest>=7.4.0
colorama>=0.4.6